langchain-test/pyproject.toml
[project]
name = "langchain-mcp-test"
version = "0.1.0"
description = "Langchain MCP Test"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "isort>=6.0.1",
    "python-dotenv>=1.1.1",
    "google-generativeai>=0.4.1",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "streamlit>=1.49.1",
]
